Add explicit initialization process
Instead of implicitly triggering a Google OAuth process if the user credential file is not found, this update provides an explicit 'init' command, which configures encryption, performs Google OAuth, and saves the data in a named 'creds' file.

Warning: The creds file format is expected to change. The stored format is versioned, but versions will only be honored in tagged releases (not during development).